[#741] step 2 remove event API and use EventTransformers
