+++ src/features/Calendars/services/deleteEventInstanceAsync.ts
@@ -1,9 +1,10 @@
-import { deleteEventInstance } from '@/features/Events/EventApi'
 import { CalendarEvent } from '@/features/Events/EventsTypes'
 import { formatReduxError } from '@/utils/errorUtils'
 import { createAsyncThunk } from '@reduxjs/toolkit'
 import { RejectedError } from '../types/RejectedError'
 import { Calendar } from '../CalendarTypes'
+import { fetchAllRecurrentVevents, putEvent } from '@/features/Events/EventDao'
+import { makeDeleteEventInstanceJCal } from '@/features/Events/EventTransformers'
 
 export const deleteEventInstanceAsync = createAsyncThunk<
   { calId: string; eventId: string },
@@ -11,7 +12,10 @@ export const deleteEventInstanceAsync = createAsyncThunk<
   { rejectValue: RejectedError }
 >('calendars/delEventInstance', async ({ cal, event }, { rejectWithValue }) => {
   try {
-    await deleteEventInstance(event)
+    const vevents = await fetchAllRecurrentVevents(event)
+    const newJCal = makeDeleteEventInstanceJCal(vevents, event)
+    await putEvent(event, newJCal)
+
     return { calId: cal.id, eventId: event.uid }
   } catch (err) {
     const error = err as { response?: { status?: number } }
