+++ src/features/Calendars/services/getEventAsync.ts
@@ -1,5 +1,6 @@
-import { getEvent } from '@/features/Events/EventApi'
+import { fetchEvent } from '@/features/Events/EventDao'
 import { CalendarEvent } from '@/features/Events/EventsTypes'
+import { parseFetchedEvent } from '@/features/Events/EventTransformers'
 import { formatReduxError } from '@/utils/errorUtils'
 import { createAsyncThunk } from '@reduxjs/toolkit'
 import { RejectedError } from '../types/RejectedError'
@@ -10,10 +11,12 @@ export const getEventAsync = createAsyncThunk<
   { rejectValue: RejectedError }
 >('calendars/getEvent', async (event, { rejectWithValue }) => {
   try {
-    const response: CalendarEvent = await getEvent(event)
+    const response = await fetchEvent(event)
+    const fetchedEvent = parseFetchedEvent(event, response)
+
     return {
       calId: event.calId,
-      event: response
+      event: fetchedEvent
     }
   } catch (err) {
     const error = err as { response?: { status?: number } }
