+++ src/features/Calendars/services/importEventFromFileAsync.ts
@@ -1,4 +1,4 @@
-import { importEventFromFile } from '@/features/Events/EventApi'
+import { importEvent } from '@/features/Events/EventDao'
 import { importFile } from '@/utils/apiUtils'
 import { formatReduxError } from '@/utils/errorUtils'
 import { createAsyncThunk } from '@reduxjs/toolkit'
@@ -21,7 +21,7 @@ export const importEventFromFileAsync = createAsyncThunk<
         status: undefined
       })
     }
-    await importEventFromFile(id, calLink)
+    await importEvent(id, calLink)
   } catch (err) {
     const error = err as { response?: { status?: number } }
     return rejectWithValue({
