+++ src/features/Calendars/services/putEventAsync.ts
@@ -1,5 +1,6 @@
-import { putEvent } from '@/features/Events/EventApi'
+import { putEvent } from '@/features/Events/EventDao'
 import { CalendarEvent } from '@/features/Events/EventsTypes'
+import { calendarEventToJCal } from '@/features/Events/utils'
 import { toRejectedError } from '@/utils/errorUtils'
 import { createAsyncThunk } from '@reduxjs/toolkit'
 import { Calendar } from '../CalendarTypes'
@@ -13,10 +14,11 @@ export const putEventAsync = createAsyncThunk<
   'calendars/putEvent',
   async ({ cal, newEvent, calType }, { rejectWithValue }) => {
     try {
-      await putEvent(
+      const jCal = calendarEventToJCal(
         newEvent,
         cal.owner?.emails ? cal.owner.emails[0] : undefined
       )
+      await putEvent(newEvent, jCal)
 
       return {
         calId: cal.id,
