+++ src/features/Events/EventPreview/EventPreviewHeader.tsx
@@ -6,7 +6,7 @@ import EditIcon from '@mui/icons-material/Edit'
 import FileDownloadOutlinedIcon from '@mui/icons-material/FileDownloadOutlined'
 import MoreVertIcon from '@mui/icons-material/MoreVert'
 import { useI18n } from 'twake-i18n'
-import { dlEvent } from '../EventApi'
+import { fetchEventIcs } from '../EventDao'
 import { CalendarEvent } from '../EventsTypes'
 
 interface EventPreviewHeaderProps {
@@ -45,7 +45,7 @@ export function EventPreviewHeader({
   const handleDownload = async (): Promise<void> => {
     let url: string | null = null
     try {
-      const icsContent = await dlEvent(event)
+      const icsContent = await fetchEventIcs(event)
       const blob = new Blob([icsContent], { type: 'text/calendar' })
       url = URL.createObjectURL(blob)
       const link = document.createElement('a')
