+++ src/features/Events/hooks/useMasterEvent.ts
@@ -1,6 +1,7 @@
 import { useState, useEffect, useMemo } from 'react'
+import { fetchEvent } from '../EventDao'
 import { CalendarEvent } from '../EventsTypes'
-import { getEvent } from '../EventApi'
+import { parseFetchedEvent } from '../EventTransformers'
 
 export function useMasterEvent(
   event: CalendarEvent | null | undefined,
@@ -40,7 +41,9 @@ export function useMasterEvent(
       setIsLoadingMasterEvent(true)
       try {
         const masterEventToFetch = { ...event, uid: baseUID }
-        const fetched = await getEvent(masterEventToFetch, true)
+        const response = await fetchEvent(event)
+        const fetched = parseFetchedEvent(masterEventToFetch, response, true)
+
         if (!cancelled) setMasterEvent(fetched)
       } catch (err) {
         console.error('Failed to fetch master event:', err)
